requirements.txt
Flask==2.0.2
plotly==5.5.0
mysql-connector==2.2.9
apscheduler==3.8.1
pandas==1.3.5


